+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 ## [Unreleased]
 ### Added    
+* Instead of passing a distribution, can now add custom values as an `InVar`, which must match the number of cases. For example, `sim.addInVar(vals=[0, 1, 2, 3])`
 * `case.vars` dict to reference all invars and outvars, and `case.vals` dict to reference all invals and outvals
 * `__getitem__` for cases to get a val by name, e.g. `case['valname']`
 * `__getitem__` for sims to get a case by number or a var by name, e.g. `sim[i]` or `sim['varname']`